build: initialize project structure and files
The project directory is based on STM32CubeIDE's project directory to be
able to debug the program with it's GUI debugger and the
"sumobot_mini_attach.launch" option.

Configure the makefile to add build options for all, clean, flash, and
debug. The flash option uses STM32CubeProgrammer and appropriate flags
to flash the MCU directly with "make flash". The debug option uses
ST-Util as the GDB server and connects with the board ST-Link to debug
through the CLI if preferred.

The purpose of this structure is to give the user the option to use the
STM32CubeIDE to build and debug, or use CLI options if they desire.