feat(dockerfile): add dockerfile to create build environment image
Modify the makefile so the path to the arm-none-eabi-gcc compiler binary
is handled by the makefile itself. This makes it easier to resolve the
path problem in the container instead of having to manually export
the path for the container user "arch". All the dockerfile needs to do
is make sure the arm gcc compiler is in the same path location
~/tools/arm-gnutoolchain-.../bin.

The dockerfile itself has the documentation to explain the choices and
commands being used in it. The overarching idea is to use a multi-stage
build to optimize file size and to circumvent a few issues trying to
install the "yay" AUR helper.

TODO: the pacman cache in /var/cache/pacman/pkg and the package metadata
directory in /var/lib/pacman/sync are not being properly deleted.