fix(CI): add docker user to root group to resolve permission issues
Github actions has a permission related issue when cleaning up the
docker container. The fix seems to be to add an options line to use the
"root" user in the container, however this causes path problems for the
makefile so instead, add the created user arch to root group as the
simplest solution.