fix(CI): modify Dockerfile/Makefile to work with root permissions
The initial docker image set up throws permissions errors when using the
checkout@v4 action. The solution is to run the container as --user root.
Edit the makefile GCC path from ~/tools/arm-none-eabi... to
/opt/arm-none-eabi... so when the github workflow runs as root, it's
able to compile the code with proper path to the arm-none-eabi-gcc
binary.

Edit the dockerfile to reflect the environmental change and move the ARM
GNU toolchain directory to /opt instead of ~/tools.

Modify .github/workflows/ci.yml to checkout the repo with submodules
otherwise the makefile won't be able to find the necessary CMSIS headers
when trying to compile