feat(gpio): implement barebones gpio hal
Add "gpio.c" and "gpio.h" as an abstraction layer to make working with
the low level CMSIS headers easier. This allows inevitably writing code
for the drivers much easier and more portable because the gpio.c/h files
act as a custom HAL to interface with the low level register reads and
writes.

Edit the makefile to comment out the debug target. Prefer to use the
STM32CubeIDE debugger which makes it fast and easy to inspect the bits
and registers.

Add .cache/ and compile_commands.json files to gitignore. These are
extra files that are required to make clangd work with neovim to enable
LSP autocomplete, syntax checking, etc.

Parts of main.c and led.c were refactored to work with the HAL of
gpio.c.