fix(debugger): modify debugger config settings to work with makefile
The debugger in STM32CubeIDE kept defaulting to the .elf file the IDE
would generate as the debugger target. When I tried to use it with my
own makefile compiled .elf file, the symbols wouldn't load properly and
I wasn't able to step into the functions and create breakpoints.

The issue was the debugger launch config was pointing to the wrong
binary (they have the same name).

Remove the unnecessary building and flashing when launching the debugger
because I manually build and flash using the makefile prior to launching
the debugger.