refactor(led): refactor the led driver to work with the GPIO HAL
The led driver was interacting directly with the hardware by using the
CMSIS headers to toggle the LEDs on and off. The code has now been
refactored so the led driver interacts with the gpio.c custom HAL rather
than writing bits directly to the registers themselves.

Add 3 functions to the gpio HAL to set, clear, and toggle bits in the
data register ODR for future hardware that will need to make use of it.
The --suppress-unused function flag in make cppcheck phony target was
also removed because it was throwing ghost nofile errors.