feat(systick_delay): implement accurate hardware timer delay
Using for-loops and software delays are not accurate compared to
hardware timers. Add systick interrupt functionality to measure
accurate delays for blinking LEDs and other future uses where delays
might be necessary.

Systick is not a GPIO hardware block so it was not implemented using a
custom HAL. In order to keep it decoupled from other hardware blocks, I
implemented it using bare metal register reads and writes.