fix(gpio hal): fix logic in gpio_mode_set function to update struct
Originally the gpio_mode_set() function was only setting the bits in the
appropriate registers to set the mode of the pin. The function was not
updating the value of the attribute inside of the struct itself, so this
means later when gpio_config_compare() function is used, the config
would return the wrong value since it has not been updated.