fix(assert handler): fix recursive assert issue
The problem with trying to use the LED drivers inside of the
assert_handler is if any of the asserts fail inside of the LED driver,
then the assert_handler won't actually be able to flash the LED.

The assert handler calls the LED toggle function, but the led toggle
function triggers the failed assert, this means the assert handler would
keep calling itself until there is a stack overflow.

Also another edge case is if the led_initialize() isn't called at all.
This means the LEDs wouldn't be initialized at all to flash when the
assert triggers on a failed or not-called led_initialize.

By initializing the LEDs bare metal inside of the assert handler, this
guarantees that regardless of if there are failed asserts in the led
driver, the assert_handler will still do its job of blinking the LEDs
visually to show a failed assertion.