fix(makefile): add extra compilation flags to optimize final executable
Add --specs=nano.specs to use the nano version of the newlib C library
to further reduce the size of the final executable.

Add floating point unit compilation flags to enable FPU hardware blocks
for future use in ADC and sensor logic arithmetic.

Add "make size" as a command to check the size of the executable to
compare how adding certain functions and compilation flags changes the
file size.