feat(uart): improve uart functionality from polling to interrupts
Replace the uart2_send_char() function in uart.c with a
uart2_interrupt_send_char() function to use interrupt based uart instead
of polling to improve efficiency.

Implement a software ring buffer to hold the characters to be sent out
while the CPU is executing other instructions. This acts as a buffer to
be used to transmit out the characters when the transmit (tx) interrupt
actually fires.