feat(printf): implement mpaland's small printf functionality for tracing
Modify uart.c and uart.h files to change the uart_send_char() function
to retarget to mpaland's printf() function.

Edit the makefile to account for new external/printf/ directory as a
build target when compiling the project.

Test in the main.c file to confirm the printf() function can print
strings and integers using the traditional stdlibc printf(%d) flag.

Add a .clangd file to suppress missing headers because otherwise, it
kept throwing a warning saying the "printf.h" file wasn't being used
directly when it clearly was. I suspect this is an issue with the LSP or
the compile_commands.json file not being configured properly even
through the use of bear.

Re-compile the compile_commands.json file using bear -- make to account
for the addition of the printf external directory.