feat(trace): implement tracing functionality
Wrap printf and uart functionality inside of a TRACE macro to be able to print
the file and line where the assert is triggered. This adds another layer
of debugging to recognize where an assert is triggered inside of the
program.

Add a trace.c/.h file that uses a TRACE macro to print the file and line
number. Use TRACE_DISABLE as a define to easily disable/enable traces in
the build.

Add a separate uart2_initialize_assert() function that doesn't include
asserts inside of the initialize function. This is to decouple it from
the uart and gpio modules and this function manually sets the registers
without any helper functions. This also prevents any possibility of
recursive assert calls.