feat(gpio_interrupt): configure EXTI gpio interrupt
Fix bug with setting SYSCFG->EXTICR register. Turns out you need to also
set the SYSCFG clock register in the RCC->APB2ENR.

Designate pin PB0 as the IR Receiver GPIO pin and configure it to
trigger an external interrupt on a falling edge. The function
gpio_interrupt_set() handles the logic of setting the correct bits in
the corresponding registers to set a GPIO Port/Pin as an external
interrupt.

The IR Receiver file handles the initialization and later, getting the
command from the IR remote. Using the logic analyzer with the protocol
decoder makes it easy to extract out the command bits.

Use LED Orange to test the EXTI0 ISR to make sure it works properly.
Interestingly, with EXTI ISR's, the interrupt software trigger flag must
be cleared manually in the EXTI->PR register by writing a 1 to the
corresponding pin bit.

Edit Makefile to include --check-level=exhaustive flag to suppress
warning. Even though it's set to exhaustive, running cppcheck is still
just as fast as before with no slowdown.

TODO: Implement timer and NEC signal decoding logic