feat(timer): refactor ir receiver code to work with a timer peripheral
Modify gpio.c IR RECEIVER pin to PA15 because this is where the timer 2
peripheral is enabled. Instead of using a GPIO Pin interrupt and then
reading that interrupt from a timer peripheral, simplify the code so the
IR Receiver outputs directly to the timer pin and the timer triggers an
interrupt on a falling edge.

TODO: implement logic to decode the incoming IR signal