feat (ir decode): decode ir signal with a timer and input capture mode
Implement a timer2 ISR to trigger on every falling edge and calculate
the pulse width time by taking the difference of the previous and
current read CCR1 value.

The time ranges for each pulse width classification is documented in the
ISR itself.

The code was tested by checking the signal values of different buttons
on the IR receiver, by first confirming with a logic analyzer and then
checking in the debugger to compare message bit values. Using a debug
trap by dumping "current" and "time_diff" values and using an if
statement to trigger a breakpoint at a given index (in this case, a
specific pulse width) I wanted to check helped a ton.

There was a weird bug where the values read from CCR1 seemed to have
been multiplied by 160 (which is the value of the timer prescaler).
After loads of debugging, the issue was resolved by changing the ARR
value from 0xFFFF.FFFF - 1 (32 bit max number) to 0xFFFF - 1 (16 bit max
value). I'm not sure why this was a problem because timer 2 is supposed
to be a 32 bit counter timer.

TODO: implement application code that can extract the command from the
command ring buffer for each button press on the remote.
Also clean up a lot of the #if0 endifs that were used to comment out
code while testing.