feat(ir receiver): implement function to use decoded command in main
Add an enum that maps all 21 buttons of the IR remote to a specific
command 8 bit pattern. This is used in conjunction with the command
value that is decoded and put into the command ring buffer.

A separate function will then check periodically without polling if the
ring buffer is not empty and then convert that to a string to print out
and test if the buttons correctly match to the printed command over
UART.