feat(adc): implement adc initialize function
Test to make sure the adc initialize function is working properly and
the respective GPIO pins for the 4 ADC channels are being set without
triggering any asserts.

Create 4 new pins inside of the gpio_board_pins[] array to represent the
4 new ADC3 Channel pins, also add UART2_RX as a new pin on the board even
if it's not being used as a completion pair for the UART2_TX pin. This
is to reserve PA3 pin in the future.

The reason ADC3 was chosen and Port F is because of possible signal
integrity issue for the ADC's if used too close to the UART2 PA2 and PA3
pins.