feat(adc): implement polling based conversion and reading values
Port F is not available as a breakout pin on the board so I had to
change the ADC channels to be used for ADC123 channels 10, 11, 12, 13
which are mapped to PC0, 1, 2, 3.

Create a new set of files called "test.c" and "test.h" which will be
used to hold scratch functions to quickly test code without having to
delete and modify the code directly in main.c. Should've done this from
the start.

The code is tested by saving the read ADC values into an array and
printing those values out using our TRACE function over UART. Each ADC
pin is floating right now so use a jumper wire to connect it to GND and
then to 3.3V to check if the ADC is properly reading in the values
manually.