feat(adc_dma): refactor ADC code to work with DMA
In order to improve efficiency, use DMA to offload memory transfer
between the ADC data register and the adc_dma_buffer from the CPU to the
DMA module.

Code was tested by printing the values using the TRACE macro onto UART
and doing the same test as before, manually pulling each pin up or down
with a jumper wire and checking the values inside of the adc_dma_buffer
to confirm they are being updated.

TODO: Test with a potentiometer and implement DMA interrupt