feat(adc_dma interrupt): add DMA interrupt to make ADC data more robust
The DMA interrupt is used to avoid directly touching/reading from the
adc_dma_buffer[] array that will be written by the DMA controller. If we
read from adc_dma_buffer we risk accessing it at the same time as the
DMA. The DMA ISR will copy the values from the adc_dma_buffer[] array to
adc_dma_buffer_safe[] array which decouples reading the data from the
DMA. This means accessing the adc_dma_buffer_safe[] is guaranteed to be
safe as long as interrupts are briefly disabled when accessing it.

Modify the NVIC Priority for the different interrupts happening in the
drivers so the priority goes like so:
ADC > IR Receiver > UART

UART has the lowest priority because printing values over UART is the
least time critical task.

Add *.pdf files to gitignore so reference manuals and data sheets don't
get tracked by git.

TODO: fix bug with being perpetually stuck in the DMA ISR. I suspect
it's because of DMA circular mode the DMA keeps flagging the TCIF
interrupt in the background even though I clear it and its being choked
up in the ISR.