feat(adc_dma_interrupt): refactor code to use single conversion dma mode
Refactor the code to use single conversion so the ADC and DMA transfer
will stop as soon as the ISR is being triggered. This makes transferring
the ADC values from adc_dma_buffer to adc_dma_buffer_safe in the ISR
without having the chance the DMA will overwrite the values in
adc_dma_buffer inside of the ISR.

There was a weird bug where the ADC DMA sampling would only happen once
and never happen again. This was resolve by having to clear the
ADC_CR2_DMA bit explicitly to 0, and then rewriting it to 1 in software
because the hardware does not clear it by itself. The reference manual says
this in ADC Section 13.8.1.

To use ADC single conversion mode with DMA interrupts, you have to clear
and write the ADC_CR2_DMA bit, explicitly reset the reload value in DMA_SxNDTR,
reenable the DMA_SxCR_EN stream, and then start the ADC conversion again
by writing a 1 to the ADC_CR2_SWSTART bit.

The refactored code has been tested by pausing and inspecting the debugger by
hooking up a potentiometer to each of the 4 ADC channels and checking if
the values were being shown in the adc_dma_buffer_safe array. Also by
checking to see if the LED GPIO pin inside the ISR was being toggled repeatedly
in the logic analyzer.

TODO: Need to test to see if the values can be successfully printed in
the TRACE function. Previously, calling the TRACE function to print the
array values was causing problems.