feat(adc_dma_interrupt): resolve unable to print adc values to terminal
Weird bug where trying to use the TRACE macro to print the ADC values
inside of adc_dma_buffer_safe array was causing issues and the DMA
wasn't working at all.

Turns out trying to use the TRACE macro inside of a critical section
that disables only the DMA transfer complete interrupt completely breaks
the DMA functionality. The workaround to this is to copy the values from
adc_dma_buffer_safe to another buffer adc_dma_print_values array inside
of a critical section disabling GLOBAL interrupts with __disable_irq().

The trace macro will then be called to simply print the values inside of
adc_dma_print_values[] without having to be called inside of a critical
section. I'm not sure entirely why this works, but it does. When
actually working with the code in the application layer, the
adc_dma_buffer_safe array should be used to access the values, NOT the
adc_dma_print_values. The print_values array is solely for
debugging/testing purposes and using values from this array is
inefficient since it's an extra unnecessary step to copy values one more
time.

TODO: implement sensor code for the QRE1113