refactor(adc_dma): redo logic for getting values from the ADC
The adc_dma_get_values() function should be used to extract the values
from the edge detecting sensor. This should be done by abstracting away
all the logic inside of the ADC driver. By having the get_values()
function take an input array and copy the values into it inside of a
critical section, the edge detecting module can be free to use those
values without having to worry about interrupt handling of the DMA
interrupt.

This decouples the code as much as possible and all the sensor needs to
do is get the adc values, and use those values inside of its own module.

TODO: code the logic for the QRE1113 sensor by abstracting away the
channels and representing the values it gets from the ADC as
locations/placements of the sensors such as front left, back right, etc.