comments(trace): add clarifying comments to the TRACE macro
Detail the syntax for the TRACE macro and how it exactly works with the
trace() function because it looks a bit confusing at first. Add comments
in both the trace.h and trace.c files.

Edit TRACE macro to not implicitly add the "\n" at the end to make it
match the behavior of traditional printf().