feat(qre1113): add drivers for the qre1113 edge detecting sensor
This layer in the software architecture is going to be used to abstract
away the ADC because the application shouldn't need to care about what
channels or streams are being used to acquire the voltages. The QRE1113
sensors will let the application know, the sensors at front, right,
back, etc LOCATIONS are the ones that are reading a certain voltage.

TODO: Solder one of the edge detecting sensors with headers and test to
make sure the readings are correct.