feat(edge_detect): implement application layer code for edge detection
The sumobot will be moving around on the dohyo trying to push the enemy
sumobot off. The white border around the dohyo signifies the edge of the
ring and the robot should know to try to get back to the middle and not
fall off when it detects a white edge on the IR sensors.

In order to separate the modules, the software stack looks like so:
edge detection application layer -> QRE1113 sensor -> ADC driver

With this architecture in mind, if the sensor or the microcontroller has
to be swapped out in the future, it decouples the code as much as
possible into modular layers that can be swapped out.

In order to keep the code clean and avoid giant nested if else
statements, I implement bitmasking and a lookup table to return where
the white edge has been detected. The software logic can be found in the
edge_detect.c file. The downside is having to use more memory for the
array, but its an elegant solution.

Also weirdly enough, in the testing print function, the edge_string
variable kept getting flagged by cppcheck as an unreadVariable
assignment for each switch statement case. The only way to silence it
was to suppress the warning for each line.