refactor(gpio): add additional gpio pin attribute called "OTYPE"
Add OTYPE attribute because it is necessary to configure I2C SDA and SCL
as output open-drain. This is because I2C can only actively drive the
pin low whereas with LEDs the output can be actively driven high or low.
Make this explicit in the code and add the corresponding checks and set
function inside of gpio.c

This also means the ASSERT() checks and gpio_config_compare() function
had to be expanded to include these changes. All other initialize()
functions inside of peripherals also need to have been edited to reflect
the changes. The code was tested and no ASSERTs are triggered.