feat(i2c): implement i2c write function and test function
Implement the i2c write function and then test both i2c read/write
within a self contained i2c_test_read_write() function that can be
called independently in main to confirm the functionality of i2c.

There is a weird issue when trying to use smaller helper functions
inside of the i2c_read() and i2c_write() functions where the program
just hangs and it doesn't work properly. Currently the implementation of
both of these functions are explicitly writing out each step of the code
which means there is a little bit of duplicate code. For the current
purpose of functionality and making sure it works, I think this is
acceptable.

Interestingly enough, the I2C lines seem to work properly even when
using the internal pull up resistors which are rated at around 40k ohms.
The general consensus seems to be using 4.7k - 10k ohm external pullup
resistors. I also did test with 1.5k ohm external pullups and it also
worked. Although it currently works with the internal pull ups, this may
need to be adjusted later down the line when multiple VL53L0X sensors
will be used because the bus capacitance will be higher.