feat(gpio): add configuration for gpio output speed attribute
The gpio output speed for protocols like I2C is important so the rise
and fall times (signal transitions) is extremely important. If the speed
is too low, it could result in timing issues and in the case of I2C
where its very sensitive to the Trise and Tfall timings, this is
crucial. Since the SCL is the main driver of the signal timings, too low
of a GPIO speed could result in the actual switching of the SCL clock
not happening fast enough. Remedy this by setting the GPIO speed to HIGH
for I2C and medium for UART/NEC protocol (ir receiver).

For LEDs, the timing is not important so leave it at 0 to conserve
power.