fix(i2c): fix bug where i2c read was hanging indefinitely
Lots of debugging in this commit. First implement error checking codes
to figure out where the problem in the i2c line was. Edit the i2c
functions to return an enum value that holds the codes for different
possible errors. This lead to uncovering the i2c kept timing out
randomly in the test function when there was a i2c_write() -> immediately
followed by a i2c_read().

Originally the fix was to add a very small delay of 1ms between the
function calls, but turns out there was a problem where the start
condition would immediately generate a stop condition and thus resulting
in the time out because the device address would never get sent and the
ADDR bit never being set.

Currently, the code works by setting the GPIO speed for the I2C SCL and
SDA pins to low, not high.

TODO: test and figure out why multi-byte i2c_read() is not working.
Multi-byte i2c_write() seems to be working fine.