fix(i2c): fix weird hanging issue with a consecutive write -> read
The issue was the wrong polling guard condition being used in the while
loop of the stop_condition() function. Originally the code was
while(!(CR2 & BUSY)) but it should've been while(CR2 & BUSY). We need to
wait until the BUSY flag is CLEARED by hardware. Not SET by hardware.
Silly mistake.

Test functionality of reading/writing 2 bytes with the i2c functions. It
seems like the write register I am testing with is a 1 byte-size
register. Logic analyzer confirms 2-byte write and reads are
functioning properly.

Confirm functionality of reading/write 4 bytes as well with the logic
analyzer and the uart debugging through serial terminal.