fix(vl53l0x): debugging vl53l0x_initialize() function
Test data_initialize() function and confirm functionality.
Test static_initialize() function and confirm functionality.

Problem in vl53l0x_single_ref_calibration() function. The code was
getting permanently stuck in the do-while loop while waiting for the
interrupt to trigger. The condition of interrupt_status & 0x07 was
forever true, because interrupt status would keep reading as 0x40 and
that means 0x40 & 0x07 == 0 since the value wasn't being read properly.

This gave the hint something must be wrong with the vl53l0x_interrupt_enable()
function. When inspected with the logic analyzer, the writes to the
registers VL53L0X_SYSTEM_INTERRUPT_CONFIG_GPIO_REG and
VL53L0X_SYSTEM_INTERUPT_CLEAR_REG were incorrect.

The program was writing a 0xC9 to SYSTEM_INTERRUPT_CONFIG_GPIO when it
should've been a 0x04, and a 0x00 to SYSTEM_INTERRUPT_CLEAR_REG when it
should've wrote a 0x01.

The solution was to not pass in casted constants. Initially I passed in
the values as such (uint8_t*)0x04 and (uint8_t*)0x01. This was bad
practice because it wasn't creating variables that held those desired
values of 0x04 and 0x01. Instead, it was just directly treating them as
addresses then derefencing them to write whatever garbage values those
0x04 and 0x01 addresses held. Instead, I declared new variables "data"
and "clear_interrupt" and set those as 0x04 and 0x01 then passed in the
addresses to these variables. This way inside of the i2c functions, they
would get properly dereferenced as the correct values.