feat(vl53l0x): implement singular range finding measurement
Using Artful Bytes blogpost as a base, modify the working code to work
with my own i2c implementation for the STM32 microcontroller.

The accuracy of the sensor is something that needs to be worked but as
of right now, the range finding is working and the sensor is able to
correctly identify that an object in range is getting closer and
further. Further testing is required but the sensor is working.

These vl53l0x breakout boards were bought from Aliexpress and 3 of them
are straight up faulty. 2 of them work within the intended range, and
the last one seems to have a slightly shorter range than the 2 that work
properly. Learned my lesson ordering from Aliexpress, will have to live
with this for now.