refactor(vl53l0x): modify code to prepare for multi vl53l0x support
The problem with trying to use multiple vl53l0x sensors is the I2C
address conflict. The refactoring will make it easier to work with
separate slave addresses because now, instead of sending the default
VL53L0X_DEVICE_ADDRESS to every single i2c_write and read functions
within the driver (hardcoding it), it's now a global variable that can
be set with the vl53l0x_set_slave_address() function prior to calling
any i2c_write() and i2c_read() function.

This makes the code more modular and will make it easy to dynamically
choose which of the 3 vl53l0x sensors the program will interact with.