refactor(vl53l0x): continue refactoring to work multiple sensors
Add second sensor functionality and test the code to make sure it works
with two separate VL53L0X sensors. Currently it works properly with just
two sensors hooked up.

All the asserts in the initialization functions like data_initialize()
and static_initialize() had to be removed because before, we were only
working with one sensor so the functions would only ever be called once.
Now we have to call it individually for each sensor so it doesn't make
sense to have a global ASSERT to check them just once at a time.

Refactor the test function to be able to get a single range measurement
while using the index of the sensor as the input parameter to call it.

Remove the unnecessary .sensor_initialized property to the struct
because it is redundant.