feat(vl53l0x): multiple sensor integration
Successfully integrate 3 VL53L0X sensors using polling and single range
measurements. As of right now to keep it simple I use polling in the main
loop to acquire the range measurements because the processor isn't doing
anything else during the downtime. All the other peripherals are using
interrupt based so polling with the LIDAR sensors should be fine.

Initially the third sensor wasn't successfully initializing and turns
out it was a problem with the pull up resistors I was using. I think due
to the bus impedence at 4.7k ohms, it was too strong so I had to half
the resistance of them down to 2.2k ohm resistors and now all 3 sensors
seem to work properly.