feat(tb6612fng): implement motor driver functionality
Use the pwm low level driver to write higher level firmware to control
the direction and speed of both of the motors using a motor driver.

Configure additional GPIO pins to use as control pins for the motor
driver. The STBY (standby) pin is wired directly from the VCC of the
microcontroller and this seems to be fine.

Test to make sure the code works properly however there is one slight
issue. At lower pwm duty cycles, the motor seems to shake and vibrate a
lot. My guess is because the PWM frequency is relatively low at just
1kHz.

TODO: Increase PWM frequency to >= 10kHz