fix(tb6612fng): increase PWM frequency and retest motors
Refactor the tb6612fng_direction() function to be cleaner and so it
doesn't have the unnecessary else statement. I can achieve the same
effect of making sure the motors run in the same direction by making
sure the polarity of the motor terminals are proper on the motor driver.

Increase the PWM frequency from 1kHz to 20kHz to smoothen out DC motor
behavior.