refactor(tb6612fng): rename variables to be more sensible
The tb6612fng motor driver file shouldn't really understand the idea of
"speeds" and "directions" for the motors. The motor driver in and of
itself really only knows to set the duty cycle for the PWM and setting
the logic for the motor control pins to move them in a specific
direction clockwise or counterclockwise. I edited the variable and
function names to better reflect this.

The concept of "speed" and "direction" should be expressed in the
application level layer code, not in the firmware for the motor driver.