feat(move): add in application layer code to move the robot
Start implementing the "movement" portion of the robot code. This is the
start of the application layer of the code and will use th TB6612FNG
motor driver firmware code for higher level logic to interface with it
easier.

Add a TODO section in the assert_handler because when an ASSERT
triggers, it should immediately STOP the motors from running.

Refactor the test code very slightly just to make it more intuitive to
call and edit the ir_receiver_cmd_get() function to be a proper public
function to be used with the motor driver, NOT as a test function in and
of itself.