feat(move): implement ir remote control functionality
Implement an application level layer code that abstracts away the PWM
and motor driver to control the robot using "human commands" such as
forward or reverse direction, and speed levels like stop, slow, fast,
etc. This makes it easier to interface with the lower level hardware.

Use the ir_receiver_get_cmd() to test the functionality of this
application level layer code. The remote control sends the desired
command to set the speed and direction of the motors.

TODO: Need to test the microcontroller with an external PSU (LiPo
batteries) and a power switch and then build a 3D modelled chassis to
prototype that it works remotely.